Реализован метод создания записей в базу данных

+++ MetricsManager/MetricsManager/DAL/Migrations/FirstMigrations.cs
@@ -26,27 +26,27 @@ namespace MetricsManager.DAL.Migrations
                   .WithColumn("AgentAddress").AsString();
             Create.Table("cpumetrics")
                   .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                  .WithColumn("AgentId").AsInt64()
+                  .WithColumn("AgentId").AsInt32()
                   .WithColumn("Value").AsInt32()
                   .WithColumn("Time").AsInt64();
             Create.Table("dotnetmetrics")
                   .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                  .WithColumn("AgentId").AsInt64()
+                  .WithColumn("AgentId").AsInt32()
                   .WithColumn("Value").AsInt32()
                   .WithColumn("Time").AsInt64();
             Create.Table("hddmetrics")
                   .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                  .WithColumn("AgentId").AsInt64()
+                  .WithColumn("AgentId").AsInt32()
                   .WithColumn("Value").AsInt32()
                   .WithColumn("Time").AsInt64();
             Create.Table("networkmetrics")
                   .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                  .WithColumn("AgentId").AsInt64()
+                  .WithColumn("AgentId").AsInt32()
                   .WithColumn("Value").AsInt32()
                   .WithColumn("Time").AsInt64();
             Create.Table("rammetrics")
                   .WithColumn("Id").AsInt64().PrimaryKey().Identity()
-                  .WithColumn("AgentId").AsInt64()
+                  .WithColumn("AgentId").AsInt32()
                   .WithColumn("Value").AsInt32()
                   .WithColumn("Time").AsInt64();
         }
